+++ projects/swimlane/ngx-ui/src/lib/components/tip/tip-status.enum.ts
@@ -1,5 +1,6 @@
 export enum TipStatus {
   Success = 'success',
   Error = 'error',
-  Notice = 'notice'
+  Notice = 'notice',
+  Warning = 'warning'
 }