+++ projects/swimlane/ngx-ui/src/lib/components/tip/tip.component.ts
@@ -12,7 +12,8 @@ import { TipStatus } from './tip-status.enum';
     class: 'ngx-tip',
     '[class.ngx-tip--success]': 'status === TipStatus.Success',
     '[class.ngx-tip--error]': 'status === TipStatus.Error',
-    '[class.ngx-tip--notice]': 'status === TipStatus.Notice'
+    '[class.ngx-tip--notice]': 'status === TipStatus.Notice',
+    '[class.ngx-tip--warning]': 'status === TipStatus.Warning'
   }
 })
 export class TipComponent {
@@ -20,13 +21,16 @@ export class TipComponent {
   status: TipStatus;
   @Input()
   isCloseable: boolean = false;
+  @Input()
+  icon: string;
   @Output()
   close = new EventEmitter();
-  icon: string;
   readonly TipStatus = TipStatus;
 
   ngOnInit() {
-    this.icon = this.status === TipStatus.Error ? 'warning-filled-sm' : 'info-filled-small';
+    if (!this.icon) {
+      this.icon = this.getIcon(this.status);
+    }
   }
 
   ngOnDestroy() {
@@ -35,5 +39,14 @@ export class TipComponent {
 
   onClose() {
     this.close.emit();
+  }
+
+  private getIcon(status: TipStatus): string {
+    const icons = {
+      [TipStatus.Error]: 'warning-filled-sm',
+      [TipStatus.Warning]: 'alert',
+      default: 'info-filled-small'
+    };
+    return icons[status] || icons['default'];
   }
 }